backend/requirements.txt
celery
redis
django_celery_beat
djangorestframework